Parse description
