+++ source/features/parse-backticks.tsx
@@ -1,28 +1,8 @@
 import './parse-backticks.css';
 import select from 'select-dom';
-import React from 'dom-chef';
 import features from '../libs/features';
 import getTextNodes from '../libs/get-text-nodes';
-
-const splittingRegex = /`(.*?)`/g;
-
-function splitTextReducer(fragment: DocumentFragment, text: string, index: number): DocumentFragment {
-	// Code is always in odd positions
-	if (index % 2 && text.length >= 1) {
-		// `span.sr-only` keeps the backticks copy-pastable but invisible
-		fragment.append(
-			<code className="rgh-parse-backticks">
-				<span className="sr-only">`</span>
-				{text}
-				<span className="sr-only">`</span>
-			</code>
-		);
-	} else if (text.length > 0) {
-		fragment.append(text);
-	}
-
-	return fragment;
-}
+import parseBackticks from '../libs/parse-backticks';
 
 function init(): void {
 	for (const title of select.all(`
@@ -30,9 +10,7 @@ function init(): void {
 		.commit-title .js-navigation-open
 	`)) {
 		for (const node of getTextNodes(title)) {
-			const fragment = node.textContent!
-				.split(splittingRegex)
-				.reduce(splitTextReducer, new DocumentFragment());
+			const fragment = parseBackticks(node.textContent!);
 
 			if (fragment.children.length > 0) {
 				node.replaceWith(fragment);
